feat: add initial configs
