+++ .env
@@ -1,0 +1,2 @@
+PORT=3000
+HTTP_PORT=8181
