+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+  "ignore": ["node_modules"],
+  "watch": ["./src/**/*"],
+  "exec": "node --loader ts-node/esm ./index.ts",
+  "ext": "ts"
+}
